+++ src/main/java/com/daml/extensions/damlmavenplugin/CodeGen.java
@@ -4,6 +4,18 @@
  */
 package com.daml.extensions.damlmavenplugin;
 
+import java.io.IOException;
+import java.lang.invoke.MethodHandle;
+import java.lang.invoke.MethodHandles;
+import java.lang.invoke.MethodType;
+import java.net.MalformedURLException;
+import java.net.URL;
+import java.net.URLClassLoader;
+import java.nio.file.Path;
+import java.util.List;
+import java.util.Optional;
+import java.util.Scanner;
+
 import org.apache.maven.artifact.Artifact;
 import org.apache.maven.artifact.repository.ArtifactRepository;
 import org.apache.maven.execution.MavenSession;
@@ -19,23 +31,15 @@ import org.apache.maven.project.ProjectBuildingRequest;
 import org.apache.maven.shared.transfer.artifact.DefaultArtifactCoordinate;
 import org.apache.maven.shared.transfer.artifact.resolve.ArtifactResolver;
 import org.apache.maven.shared.transfer.artifact.resolve.ArtifactResolverException;
-import java.io.IOException;
-import java.lang.invoke.MethodHandle;
-import java.lang.invoke.MethodHandles;
-import java.lang.invoke.MethodType;
-import java.net.MalformedURLException;
-import java.net.URL;
-import java.net.URLClassLoader;
-import java.nio.file.Paths;
-import java.util.List;
-import java.util.Optional;
-import java.util.Scanner;
 
 @Mojo(name = "codegen", defaultPhase = LifecyclePhase.GENERATE_SOURCES)
 public class CodeGen extends MojoBase {
 
     @Parameter(defaultValue = "${project.build.directory}/${project.artifactId}.dar")
     private String darName;
+
+    @Parameter(defaultValue = "${project.baseDir}")
+    private String damlYamlDir;
 
     @Parameter(defaultValue = "daml.Decorder")
     private String decoderClass;
@@ -115,6 +119,7 @@ public class CodeGen extends MojoBase {
         return new URLClassLoader(new URL[] { cp });
     }
 
+    @SuppressWarnings("unchecked")
     private String findJavaBindingsVersion() throws MojoFailureException {
         Optional<Artifact> codegenOpt = project
                 .getDependencyArtifacts()
@@ -128,10 +133,12 @@ public class CodeGen extends MojoBase {
         }
     }
 
-    private static String getDamlVersion() throws MojoFailureException {
+    private String getDamlVersion() throws MojoFailureException {
         String errorMsg =
                 "Cannot determine project sdk version. Make sure that `daml.yaml` includes a line specifying `sdk-version`.";
-        try (Scanner scanner = new Scanner(Paths.get("daml.yaml"))) {
+
+        Path damlYaml = getDamlYamlFile();
+        try (Scanner scanner = new Scanner(damlYaml)) {
             while (scanner.hasNextLine()) {
                 String line = scanner.nextLine().trim();
                 if (line.startsWith("sdk-version:")) {